requirements.txt
# Core Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.24.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# LLM and AI
google-generativeai>=0.8.0
openai>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Vector Store and Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.27.0
tenacity>=8.2.0
tqdm>=4.66.0

# Statistics
scipy>=1.11.0
statsmodels>=0.14.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
